Another day, another ~10 small changes with Styler. Only 9, as one of these was already formatted perfectly in the Tidyverse style.
Additionally: one change to `extract` and `extract_cues` so the error "inputs for all arguments are required" is given if any of the inputs are missing. Doesn't seem necessary, though? I tried running the previous version with one variable missing and it gave me a different error that came to the same thing. I'll leave it in ¯\_(ツ)_/¯